+++ packages/plugin-gallery/web/src/gallery-component.jsx
@@ -196,48 +196,21 @@ class GalleryComponent extends PureComponent {
   };
 
   render() {
-    const {
-      settings,
-      componentData,
-      onClick,
-      className,
-      blockProps,
-      theme,
-      isMobile,
-      helpers,
-      disableRightClick,
-      setInPluginEditingMode,
-      anchorTarget,
-      relValue,
-      disabledRevealSpoilerBtn,
-      EditableSpoilerDescription,
-      t,
-      block,
-      store,
-    } = this.props;
-
     return (
       <>
         <GalleryViewer
           key={this.state.key}
-          componentData={componentData}
-          onClick={onClick}
-          className={className}
-          settings={settings}
-          theme={theme}
-          helpers={helpers}
-          disableRightClick={disableRightClick}
-          isMobile={isMobile}
-          anchorTarget={anchorTarget}
-          relValue={relValue}
+          componentData={this.props.componentData}
+          onClick={this.props.onClick}
+          className={this.props.className}
+          settings={this.props.settings}
+          theme={this.props.theme}
+          helpers={this.props.helpers}
+          disableRightClick={this.props.disableRightClick}
+          isMobile={this.props.isMobile}
+          anchorTarget={this.props.anchorTarget}
+          relValue={this.props.relValue}
           blockKey={this.blockKey}
-          disabledRevealSpoilerBtn={disabledRevealSpoilerBtn}
-          setInPluginEditingMode={setInPluginEditingMode}
-          EditableSpoilerDescription={EditableSpoilerDescription}
-          blockProps={blockProps}
-          t={t}
-          store={store}
-          block={block}
         />
         {this.state.isLoading && this.renderLoader()}
       </>
@@ -257,14 +230,10 @@ GalleryComponent.propTypes = {
   settings: PropTypes.object,
   helpers: PropTypes.object.isRequired,
   disableRightClick: PropTypes.bool,
-  disabledRevealSpoilerBtn: PropTypes.bool,
-  EditableSpoilerDescription: PropTypes.bool,
-  setInPluginEditingMode: PropTypes.func,
   theme: PropTypes.object.isRequired,
   isMobile: PropTypes.bool.isRequired,
   anchorTarget: PropTypes.string.isRequired,
   relValue: PropTypes.string.isRequired,
-  t: PropTypes.func.isRequired,
 };
 
 export { GalleryComponent as Component, DEFAULTS };